update diagrama y clases del paquete geografia

+++ src/main/java/ar/com/sge/geografia/Transformador.java
@@ -45,7 +45,7 @@ public class Transformador {
 		this.listaDeclientesConectados = new ArrayList <>(); 
 	}//constructor
 	
-	public float totalDeConsumoDelTransformadores() {
+	public double totalDeConsumoDelTransformadores() {
 		float totalDeConsumoDeCadaCliente=0;
 		for(Cliente unCliente : listaDeclientesConectados) {
 			totalDeConsumoDeCadaCliente += unCliente.consumoDeEnergia();
